# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ремонт часов на Алексеевской",
        "contact": {
            "name": "Анастасия Паршенкова",
            "email": "parshenkova.anastasya@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/": {
            "get": {
                "tags": [
                    "Product"
                ],
                "description": "Возвращает перечень товаров",
                "operationId": "a70e7d05d2d37b125697c9e7a5e57c7c",
                "responses": {
                    "200": {
                        "description": "Перечень товаров",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "description": "Создаёт запись о товаре",
                "operationId": "4a7fe0fee0b035a362bc86505d6bc08f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Запись о товаре создана",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product}/": {
            "get": {
                "tags": [
                    "Product"
                ],
                "description": "Возвращает товар по идентификатору",
                "operationId": "d412dc47e985603132178e1387b037c1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/product"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "description": "Удаляет запись о товаре",
                "operationId": "ecd2256be348d90699334e35d9395fe6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/product"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись о товаре удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Product"
                ],
                "description": "Корректирует запись о товаре",
                "operationId": "70bc9c35a32df84550fdd73f5e255c9d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/product"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись о товаре откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reviewes/": {
            "get": {
                "tags": [
                    "Review"
                ],
                "description": "Возвращает перечень отзывов",
                "operationId": "83e9f7cb340615664bc5c0666f4081fd",
                "responses": {
                    "200": {
                        "description": "Перечень отзывов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Отзыв не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "description": "Создаёт отзыв",
                "operationId": "e41c88949b628c3a2e342b9c2a6a7f53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Отзыв создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reviewes/{review}/": {
            "get": {
                "tags": [
                    "Review"
                ],
                "description": "Возвращает отзыв по идентификатору",
                "operationId": "8a299749a7e3086f087144ef68bfd8ec",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/review"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отзыв",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Отзыв не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "description": "Удаляет запись об отзыве",
                "operationId": "1d7eec5b3c17e263798f05473daea5a4",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/review"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись об отзыве удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Review"
                ],
                "description": "Корректирует запись об отзыве",
                "operationId": "390b7a877568ddc465123dbd2d8ca91f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/review"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись об отзыве откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "required": [
                    "name",
                    "photo",
                    "price",
                    "quantity",
                    "type_id"
                ],
                "properties": {
                    "name": {
                        "description": "Название продукта",
                        "type": "string"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "integer",
                        "format": "float"
                    },
                    "properties": {
                        "description": "Описание продукта",
                        "type": "string",
                        "format": "json"
                    },
                    "quantity": {
                        "description": "Количество на складе",
                        "type": "integer"
                    },
                    "available": {
                        "description": "Доступность продукта",
                        "type": "boolean"
                    },
                    "type_id": {
                        "description": "ID типа продукта",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review",
                "required": [
                    "content",
                    "rating",
                    "order_product_id"
                ],
                "properties": {
                    "content": {
                        "description": "Текст отзыва",
                        "type": "string"
                    },
                    "rating": {
                        "description": "Оценка",
                        "type": "integer"
                    },
                    "order_product_id": {
                        "description": "ID заказанного продукта",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "product": {
                "name": "product",
                "in": "path",
                "description": "Первичный ключ"
            },
            "review": {
                "name": "review",
                "in": "path",
                "description": "Первичный ключ"
            },
...
        }
    },
...
}